Added bindings to build a WASM version of the JPEG 2000 decoder
This change provides bindings in js/wasm/src/decoder.cpp that allow
the JPEG2000 decoder to be called from Javascript. It also contains
a build script to build the library and an example file
js/wasm/example.html that demonstrates use of th bindings.

Full documentation is given in js/wasm/README.md